+++ niceman/distributions/tests/test_debian.py
@@ -6,10 +6,14 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-
+import json
 import os
+from os.path import islink
+from os.path import join, isfile
 
 from pprint import pprint
+
+import attr
 
 from niceman.distributions.debian import DebTracer
 from niceman.distributions.debian import DEBPackage
@@ -19,6 +23,7 @@ import pytest
 
 import mock
 
+from niceman.support.exceptions import CommandError
 from niceman.tests.utils import skip_if_no_apt_cache
 
 
@@ -38,7 +43,7 @@ def test_dpkg_manager_identify_packages():
     distributions = list(tracer.identify_distributions(files))
     assert len(distributions) == 1
     distribution, unknown_files = distributions[0]
-    pprint(distribution)
+    print(json.dumps(attr.asdict(distribution), indent=4))
     assert distribution.apt_sources
     # Make sure both a non-local origin was found
     for o in distribution.apt_sources:
@@ -52,6 +57,34 @@ def test_dpkg_manager_identify_packages():
             break
     else:
         assert False, "A non-local origin must be found"
+
+
+@pytest.mark.slow
+@skip_if_no_apt_cache
+def test_check_bin_packages():
+    # Gather files in /usr/bin and /usr/lib
+    files = list_all_files("/usr/bin") + list_all_files("/usr/lib")
+    tracer = DebTracer()
+    distributions = list(tracer.identify_distributions(files))
+    assert len(distributions) == 1
+    distribution, unknown_files = distributions[0]
+    print(json.dumps(attr.asdict(distribution), indent=4))
+    non_local_origins = [o for o in distribution.apt_sources if o.site]
+    assert len(non_local_origins) > 0, "A non-local origin must be found"
+    for o in non_local_origins:
+        # Loop over mandatory attributes
+        for a in ["name", "component", "archive", "codename",
+                  "origin", "label", "site", "archive_uri"]:
+            assert getattr(o, a), "A non-local origin needs a " + a
+    assert len(unknown_files) == 0, "Files not found in packages: " + \
+                                    str(unknown_files)
+
+
+def list_all_files(dir):
+    files = [join(dir, f) for f in os.listdir(dir)
+             if (isfile(join(dir, f)) and
+                 not islink(join(dir, f)))]
+    return files
 
 
 # def test_find_release_file():
@@ -84,6 +117,7 @@ def test_utf8_file():
     # Simple sanity check that the pipeline works with utf-8
     (packages, unknown_files) = \
         manager.identify_packages_from_files(files)
+    packages = manager.get_details_for_packages(packages)
     # Print for manual debugging
     pprint(unknown_files)
     pprint(packages)
@@ -118,9 +152,11 @@ def test_get_packagefields_for_files():
              '/bogus'
              ]
 
-    def _run_dpkg_query(subfiles):
+    def exec_cmd_batch_mock(session, cmd, subfiles, exc_classes):
         assert subfiles == files  # we get all of the passed in
-        return """\
+        assert cmd == ['dpkg-query', '-S']
+
+        yield ("""\
 diversion by dash from: /bin/sh
 diversion by dash to: /bin/sh.distrib
 dash: /bin/sh
@@ -129,8 +165,9 @@ zlib1g:amd64: /lib/x86_64-linux-gnu/libz.so.1.2.8
 afni: /usr/lib/afni/bin/afni
 fail2ban: /usr/bin/fail2ban-server
 fail2ban: /usr/bin/fail2ban-server
-"""
-    with mock.patch.object(manager, "_run_dpkg_query", _run_dpkg_query):
+""", None, None)
+    with mock.patch('niceman.distributions.debian.execute_command_batch',
+                    exec_cmd_batch_mock):
         out = manager._get_packagefields_for_files(files)
 
     assert out == {