+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented (for humans) in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2018-01-05
+Minor release with a few fixes and performance enhancements
+### Added
+- Create apt .sources files pointing to APT snapshot repositories
+### Performance
+- Batch commands invocations in Debian tracer to significantly speed up
+  retracing
+### Fixed
+- Output of the (re)traced spec into a file
+
 ## [0.0.3] - 2017-12-19
 A minor release to demonstrate `retrace` functionality